Refactor PartnershipsTrust and TrustedPartners components to use BrandMarquee for partner logos, enhancing code reusability and maintainability.
